reorg modules + add documentations
