+++ optd-sqlplannertest/src/bench_helper.rs
@@ -1,19 +1,29 @@
-use std::{collections::VecDeque, future::Future, sync::Arc};
+pub mod execution;
+pub mod planning;
 
-use crate::{extract_flags, DatafusionDBMS, TestFlags};
+use std::future::Future;
+
+use crate::TestFlags;
 use anyhow::Result;
-use datafusion::{execution::TaskContext, physical_plan::ExecutionPlan, sql::parser::Statement};
 use tokio::runtime::Runtime;
 
+pub use execution::ExecutionBenchRunner;
+pub use planning::PlanningBenchRunner;
+
 pub trait PlannerBenchRunner {
+    /// Describes what the benchmark is evaluating.
     const BENCH_NAME: &str;
+    /// Benchmark's input.
     type BenchInput;
 
+    /// Setups the necessary environment for the benchmark based on the test case.
+    /// Returns the input needed for the benchmark.
     fn setup(
         &mut self,
         test_case: &sqlplannertest::ParsedTestCase,
     ) -> impl std::future::Future<Output = Result<(Self::BenchInput, TestFlags)>> + Send;
 
+    /// Runs the actual benchmark based on the test case and input.
     fn bench(
         self,
         input: Self::BenchInput,
@@ -22,6 +32,7 @@ pub trait PlannerBenchRunner {
     ) -> impl std::future::Future<Output = Result<()>> + Send;
 }
 
+/// Sync wrapper for [`PlannerBenchRunner::setup`]
 pub fn bench_setup<F, Ft, R>(
     runtime: &Runtime,
     runner_fn: F,
@@ -39,6 +50,7 @@ where
     })
 }
 
+/// Sync wrapper for [`PlannerBenchRunner::bench`]
 pub fn bench_run<R>(
     runtime: &Runtime,
     runner: R,
@@ -49,104 +61,4 @@ pub fn bench_run<R>(
     R: PlannerBenchRunner,
 {
     runtime.block_on(async { runner.bench(input, testcase, flags).await.unwrap() });
-}
-
-pub struct PlanningBenchRunner(DatafusionDBMS);
-
-impl PlanningBenchRunner {
-    pub async fn new() -> Result<Self> {
-        Ok(PlanningBenchRunner(DatafusionDBMS::new().await?))
-    }
-}
-
-impl PlannerBenchRunner for PlanningBenchRunner {
-    const BENCH_NAME: &str = "planning";
-    type BenchInput = VecDeque<Statement>;
-    async fn setup(
-        &mut self,
-        test_case: &sqlplannertest::ParsedTestCase,
-    ) -> Result<(Self::BenchInput, TestFlags)> {
-        for sql in &test_case.before_sql {
-            self.0.execute(sql, &TestFlags::default()).await?;
-        }
-        let bench_task = test_case
-            .tasks
-            .iter()
-            .find(|x| x.starts_with("bench"))
-            .unwrap();
-        let flags = extract_flags(bench_task)?;
-        self.0.setup(&flags).await?;
-        let statements = self.0.parse_sql(&test_case.sql).await?;
-
-        Ok((statements, flags))
-    }
-    async fn bench(
-        self,
-        input: Self::BenchInput,
-        _test_case: &sqlplannertest::ParsedTestCase,
-        flags: &TestFlags,
-    ) -> Result<()> {
-        for stmt in input {
-            self.0.create_physical_plan(stmt, flags).await?;
-        }
-        Ok(())
-    }
-}
-
-pub struct ExecutionBenchRunner {
-    pub dbms: DatafusionDBMS,
-    pub populate_sql: String,
-}
-
-impl ExecutionBenchRunner {
-    pub async fn new(populate_sql: String) -> Result<Self> {
-        Ok(ExecutionBenchRunner {
-            dbms: DatafusionDBMS::new().await?,
-            populate_sql,
-        })
-    }
-}
-
-impl PlannerBenchRunner for ExecutionBenchRunner {
-    const BENCH_NAME: &str = "execution";
-    type BenchInput = Vec<(Arc<dyn ExecutionPlan>, Arc<TaskContext>)>;
-    async fn setup(
-        &mut self,
-        test_case: &sqlplannertest::ParsedTestCase,
-    ) -> Result<(Self::BenchInput, TestFlags)> {
-        for sql in &test_case.before_sql {
-            self.dbms.execute(sql, &TestFlags::default()).await?;
-        }
-        for sql in self.populate_sql.split(";\n") {
-            self.dbms.execute(sql, &TestFlags::default()).await?;
-        }
-
-        let bench_task = test_case
-            .tasks
-            .iter()
-            .find(|x| x.starts_with("bench"))
-            .unwrap();
-        let flags = extract_flags(bench_task)?;
-
-        self.dbms.setup(&flags).await?;
-        let statements = self.dbms.parse_sql(&test_case.sql).await?;
-
-        let mut physical_plans = Vec::new();
-        for statement in statements {
-            physical_plans.push(self.dbms.create_physical_plan(statement, &flags).await?);
-        }
-
-        Ok((physical_plans, flags))
-    }
-    async fn bench(
-        self,
-        input: Self::BenchInput,
-        _test_case: &sqlplannertest::ParsedTestCase,
-        _flags: &TestFlags,
-    ) -> Result<()> {
-        for (physical_plan, task_ctx) in input {
-            self.dbms.execute_physical(physical_plan, task_ctx).await?;
-        }
-        Ok(())
-    }
 }
